Upgrade to tokio mpsc
