+++ src/sink.rs
@@ -31,6 +31,7 @@ use std::{
     pin::Pin,
     task::{Context, Poll},
 };
+use tokio::sync::mpsc;
 
 impl<T: ?Sized, Item> VecSinkExt<Item> for T where T: Sink<Item> {}
 
@@ -88,5 +89,36 @@ where
                 }
             }
         }
+    }
+}
+
+/// Wrapper for mpsc::Sender to turn it into a Sink.
+pub struct BoundedSink<T> {
+    sender: mpsc::Sender<T>,
+}
+
+impl<T> BoundedSink<T> {
+    pub fn new(sender: mpsc::Sender<T>) -> Self {
+        Self { sender }
+    }
+}
+
+impl<T> Sink<T> for BoundedSink<T> {
+    type Error = ();
+    fn poll_ready(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.sender
+            .poll_ready(cx)
+            .map_err(|error| error!(message = "Sender error.", %error))
+    }
+    fn start_send(mut self: Pin<&mut Self>, item: T) -> Result<(), Self::Error> {
+        self.sender
+            .try_send(item)
+            .map_err(|error| error!(message = "Sender error.", %error))
+    }
+    fn poll_flush(self: Pin<&mut Self>, _: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Poll::Ready(Ok(()))
+    }
+    fn poll_close(self: Pin<&mut Self>, _: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Poll::Ready(Ok(()))
     }
 }
